[REFACTOR] Move room view related files in client
