+++ pandora-client-web/src/components/wardrobe/views/wardrobeOutfitView.tsx
@@ -20,8 +20,8 @@ import { ImportDialog } from '../../exportImport/importDialog';
 import { GraphicsSceneBackgroundRenderer } from '../../../graphics/graphicsSceneRenderer';
 import { CHARACTER_PIVOT_POSITION, GraphicsCharacter } from '../../../graphics/graphicsCharacter';
 import { usePlayerState } from '../../gameContext/playerContextProvider';
-import { useChatRoomCharacterOffsets } from '../../chatroom/chatRoomCharacter';
-import { usePlayerVisionFilters } from '../../chatroom/chatRoomScene';
+import { useChatRoomCharacterOffsets } from '../../../graphics/room/roomCharacter';
+import { usePlayerVisionFilters } from '../../../graphics/room/roomScene';
 import classNames from 'classnames';
 import { useBrowserSessionStorage } from '../../../browserStorage';
