+++ pandora-client-web/src/graphics/room/roomDevice.tsx
@@ -20,26 +20,26 @@ import * as PIXI from 'pixi.js';
 import { useObservable } from '../../observable';
 import { ChildrenProps } from '../../common/reactTypes';
 import { GraphicsManagerInstance } from '../../assets/graphicsManager';
-import { CHARACTER_BASE_Y_OFFSET, CHARACTER_PIVOT_POSITION, GraphicsCharacter, PointLike } from '../../graphics/graphicsCharacter';
+import { CHARACTER_BASE_Y_OFFSET, CHARACTER_PIVOT_POSITION, GraphicsCharacter, PointLike } from '../graphicsCharacter';
 import { Container, Graphics, Sprite, useApp } from '@pixi/react';
-import { useTexture } from '../../graphics/useTexture';
-import { useDebugConfig } from './chatroomDebug';
-import { MASK_SIZE, SwapCullingDirection, useItemColor } from '../../graphics/graphicsLayer';
+import { useTexture } from '../useTexture';
+import { useDebugConfig } from '../../ui/screens/room/roomDebug';
+import { MASK_SIZE, SwapCullingDirection, useItemColor } from '../graphicsLayer';
 import { Immutable } from 'immer';
 import { useAsyncEvent, useEvent } from '../../common/useEvent';
 import _ from 'lodash';
 import { ShardConnector } from '../../networking/shardConnector';
 import { Character } from '../../character/character';
-import { useCharacterRestrictionsManager, useChatRoomCharacters } from '../gameContext/chatRoomContextProvider';
+import { useCharacterRestrictionsManager, useChatRoomCharacters } from '../../components/gameContext/gameStateContextProvider';
 import type { FederatedPointerEvent } from 'pixi.js';
 import { z } from 'zod';
 import { BrowserStorage } from '../../browserStorage';
-import { IChatRoomMode, useCharacterDisplayFilters, usePlayerVisionFilters } from './chatRoomScene';
-import { useChatRoomCharacterOffsets } from './chatRoomCharacter';
-import { RoomDeviceRenderContext } from './chatRoomDeviceContext';
-import { EvaluateCondition } from '../../graphics/utility';
-import { useStandaloneConditionEvaluator } from '../../graphics/appearanceConditionEvaluator';
-import { MovementHelperGraphics } from '../../graphics/movementHelper';
+import { IChatRoomMode, useCharacterDisplayFilters, usePlayerVisionFilters } from './roomScene';
+import { useChatRoomCharacterOffsets } from './roomCharacter';
+import { RoomDeviceRenderContext } from './roomDeviceContext';
+import { EvaluateCondition } from '../utility';
+import { useStandaloneConditionEvaluator } from '../appearanceConditionEvaluator';
+import { MovementHelperGraphics } from '../movementHelper';
 
 const PIVOT_TO_LABEL_OFFSET = 100 - CHARACTER_BASE_Y_OFFSET;
 const DEVICE_WAIT_DRAG_THRESHOLD = 400; // ms
