+++ pandora-client-web/src/graphics/room/roomScene.tsx
@@ -17,23 +17,23 @@ import React, { ReactElement, useCallback, useEffect, useMemo, useState } from '
 import { useEvent } from '../../common/useEvent';
 import { Character, useCharacterData } from '../../character/character';
 import { ShardConnector } from '../../networking/shardConnector';
-import { useChatRoomInfo, useChatRoomCharacters, useCharacterRestrictionsManager, useCharacterState, useRoomState } from '../gameContext/chatRoomContextProvider';
-import { useShardConnector } from '../gameContext/shardConnectorContextProvider';
-import { usePlayer, usePlayerState } from '../gameContext/playerContextProvider';
+import { useChatRoomInfo, useChatRoomCharacters, useCharacterRestrictionsManager, useCharacterState, useRoomState } from '../../components/gameContext/gameStateContextProvider';
+import { useShardConnector } from '../../components/gameContext/shardConnectorContextProvider';
+import { usePlayer, usePlayerState } from '../../components/gameContext/playerContextProvider';
 import { IBounceOptions } from 'pixi-viewport';
 import { CommonProps } from '../../common/reactTypes';
 import { useAssetManager } from '../../assets/assetManager';
-import { ChatroomDebugConfig, useDebugConfig } from './chatroomDebug';
-import { PixiViewportSetupCallback } from '../../graphics/pixiViewport';
-import { GraphicsBackground, GraphicsScene, GraphicsSceneProps } from '../../graphics/graphicsScene';
-import { ChatRoomCharacterInteractive } from './chatRoomCharacter';
-import { PointLike } from '../../graphics/graphicsCharacter';
-import { ChatRoomDeviceInteractive, ChatRoomDeviceMovementTool, useIsRoomConstructionModeEnabled } from './chatRoomDevice';
-import { useChatroomRequired } from '../gameContext/chatRoomContextProvider';
-import { shardConnectorContext } from '../gameContext/shardConnectorContextProvider';
+import { ChatroomDebugConfig, useDebugConfig } from '../../ui/screens/room/roomDebug';
+import { PixiViewportSetupCallback } from '../pixiViewport';
+import { GraphicsBackground, GraphicsScene, GraphicsSceneProps } from '../graphicsScene';
+import { ChatRoomCharacterInteractive } from './roomCharacter';
+import { PointLike } from '../graphicsCharacter';
+import { ChatRoomDeviceInteractive, ChatRoomDeviceMovementTool, useIsRoomConstructionModeEnabled } from './roomDevice';
+import { useChatroomRequired } from '../../components/gameContext/gameStateContextProvider';
+import { shardConnectorContext } from '../../components/gameContext/shardConnectorContextProvider';
 import { DeviceContextMenu } from './contextMenus/deviceContextMenu';
 import { CharacterContextMenu } from './contextMenus/characterContextMenu';
-import { directoryConnectorContext, useCurrentAccountSettings } from '../gameContext/directoryConnectorContextProvider';
+import { directoryConnectorContext, useCurrentAccountSettings } from '../../components/gameContext/directoryConnectorContextProvider';
 import { Immutable } from 'immer';
 
 const BONCE_OVERFLOW = 500;
