+++ pandora-client-web/src/routing/Routes.tsx
@@ -6,9 +6,9 @@ import { usePlayerData } from '../components/gameContext/playerContextProvider';
 import { Settings } from '../components/settings/settings';
 import { CharacterCreate } from '../components/characterCreate/characterCreate';
 import { CharacterSelect } from '../components/characterSelect/characterSelect';
-import { Chatroom } from '../components/chatroom/chatroom';
-import { ChatroomAdmin, ChatroomCreate } from '../components/chatroomAdmin/chatroomAdmin';
-import { ChatroomSelect } from '../components/chatroomSelect/chatroomSelect';
+import { Chatroom } from '../ui/screens/room/room';
+import { ChatroomAdmin, ChatroomCreate } from '../ui/screens/spaceConfiguration/spaceConfiguration';
+import { ChatroomSelect } from '../ui/screens/spacesSearch/spacesSearch';
 import { useCurrentAccount } from '../components/gameContext/directoryConnectorContextProvider';
 import { useShardConnector } from '../components/gameContext/shardConnectorContextProvider';
 import { AuthPage } from '../components/login/authPage';
