+++ pandora-client-web/src/ui/components/chat/chatInput.tsx
@@ -1,27 +1,26 @@
 import { AssertNotNullable, CharacterId, EMPTY_ARRAY, ChatCharacterStatus, IChatType, SpaceId, ZodTransformReadonly } from 'pandora-common';
 import React, { createContext, ForwardedRef, forwardRef, ReactElement, ReactNode, RefObject, useCallback, useContext, useEffect, useMemo, useRef, useState } from 'react';
 import { clamp } from 'lodash';
-import { Character } from '../../character/character';
-import { IMessageParseOptions, useChatroom, useChatRoomCharacters, useChatRoomMessageSender, useChatroomRequired, useChatRoomSetPlayerStatus, useChatRoomStatus } from '../gameContext/chatRoomContextProvider';
-import { useEvent } from '../../common/useEvent';
+import { Character } from '../../../character/character';
+import { IMessageParseOptions, useChatroom, useChatRoomCharacters, useChatRoomMessageSender, useChatroomRequired, useChatRoomSetPlayerStatus, useChatRoomStatus } from '../../../components/gameContext/gameStateContextProvider';
+import { useEvent } from '../../../common/useEvent';
 import { AutocompleteDisplayData, CommandAutocomplete, CommandAutocompleteCycle, COMMAND_KEY, RunCommand, ICommandInvokeContext } from './commandsProcessor';
 import { toast } from 'react-toastify';
-import { TOAST_OPTIONS_ERROR } from '../../persistentToast';
-import { Button } from '../common/button/button';
-import { usePlayerId } from '../gameContext/playerContextProvider';
-import './chatroom.scss';
-import { BrowserStorage } from '../../browserStorage';
-import { useShardConnector } from '../gameContext/shardConnectorContextProvider';
+import { TOAST_OPTIONS_ERROR } from '../../../persistentToast';
+import { Button } from '../../../components/common/button/button';
+import { usePlayerId } from '../../../components/gameContext/playerContextProvider';
+import { BrowserStorage } from '../../../browserStorage';
+import { useShardConnector } from '../../../components/gameContext/shardConnectorContextProvider';
 import classNames from 'classnames';
-import { Row } from '../common/container/container';
+import { Row } from '../../../components/common/container/container';
 import { GetChatModeDescription } from './commands';
-import { useDirectoryConnector } from '../gameContext/directoryConnectorContextProvider';
-import { Select } from '../common/select/select';
-import settingsIcon from '../../assets/icons/setting.svg';
+import { useDirectoryConnector } from '../../../components/gameContext/directoryConnectorContextProvider';
+import { Select } from '../../../components/common/select/select';
+import settingsIcon from '../../../assets/icons/setting.svg';
 import { z } from 'zod';
 import { useNavigate } from 'react-router';
-import { useNullableObservable } from '../../observable';
-import { useTextFormattingOnKeyboardEvent } from '../../common/useTextFormattingOnKeyboardEvent';
+import { useNullableObservable } from '../../../observable';
+import { useTextFormattingOnKeyboardEvent } from '../../../common/useTextFormattingOnKeyboardEvent';
 
 type Editing = {
 	target: number;
