+++ pandora-client-web/src/ui/components/chat/chatParser.ts
