+++ pandora-client-web/src/ui/screens/spaceConfiguration/spaceConfiguration.scss
