+++ pandora-client-web/src/ui/screens/spacesSearch/spacesSearch.scss
